Refactor top navigation
Refactored the top navigation into a separate file for better organization and maintainability.
