+++ src/components/navigation/TopNav.tsx
@@ -1,0 +1,83 @@
+
+import { Button } from "@/components/ui/button";
+import { Link, useLocation, useNavigate } from "react-router-dom";
+import { Menu } from "lucide-react";
+import { useState } from "react";
+
+const TopNav = () => {
+  const navigate = useNavigate();
+  const location = useLocation();
+  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
+
+  const isActive = (path: string) => location.pathname === path;
+
+  const navigationLinks = [
+    { path: "/features", label: "Features" },
+    { path: "/why-us", label: "Why Us?" },
+    { path: "/tech-stack", label: "Tech Stack" },
+  ];
+
+  return (
+    <header className="border-b relative">
+      <div className="container mx-auto px-4">
+        <div className="flex h-16 items-center justify-between">
+          <div className="flex items-center gap-8">
+            <Link to="/" className="text-xl font-bold">
+              Open Office Survey
+            </Link>
+            <nav className="hidden md:flex items-center gap-6">
+              {navigationLinks.map((link) => (
+                <Link
+                  key={link.path}
+                  to={link.path}
+                  className={`text-sm ${
+                    isActive(link.path)
+                      ? "text-primary font-medium"
+                      : "text-gray-600 hover:text-gray-900"
+                  }`}
+                >
+                  {link.label}
+                </Link>
+              ))}
+            </nav>
+          </div>
+          <div className="flex items-center gap-4">
+            <Button variant="outline" onClick={() => navigate("/login")}>
+              Login
+            </Button>
+            <button
+              className="md:hidden"
+              onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
+            >
+              <Menu className="h-6 w-6" />
+            </button>
+          </div>
+        </div>
+      </div>
+      
+      {/* Mobile Navigation Menu */}
+      {isMobileMenuOpen && (
+        <div className="md:hidden absolute top-full left-0 right-0 bg-white border-b z-50">
+          <nav className="container mx-auto px-4 py-4">
+            {navigationLinks.map((link) => (
+              <Link
+                key={link.path}
+                to={link.path}
+                className={`block py-2 text-sm ${
+                  isActive(link.path)
+                    ? "text-primary font-medium"
+                    : "text-gray-600 hover:text-gray-900"
+                }`}
+                onClick={() => setIsMobileMenuOpen(false)}
+              >
+                {link.label}
+              </Link>
+            ))}
+          </nav>
+        </div>
+      )}
+    </header>
+  );
+};
+
+export default TopNav;
